Move get_sqlalchemy_connection to its own file.
